generalize liquid and js to get fundraiser forms working
