+++ app/assets/javascripts/sumofus/backbone/action_bar.js
@@ -1,6 +1,8 @@
 const StickyMethods = require('sumofus/backbone/sticky_methods');
+const FormMethods = require('sumofus/backbone/form_methods');
 
-const ActionBar = Backbone.View.extend(_.extend(StickyMethods, {
+const ActionBar = Backbone.View.extend(_.extend(
+  StickyMethods, FormMethods, {
 
   el: '.action-bar',
 
@@ -9,7 +11,8 @@ const ActionBar = Backbone.View.extend(_.extend(StickyMethods, {
     'click .action-bar__close-button': 'hide',
     'click .action-bar__expand-arrow': 'toggleBlurb',
     'click .action-bar__top': 'toggleBlurb',
-    'click .action-bar__clear-form': 'clearForm'
+    'click .action-bar__clear-form': 'clearForm',
+    'ajax:success form.action': 'handleSuccess',
   },
 
   initialize: function() {
@@ -18,6 +21,15 @@ const ActionBar = Backbone.View.extend(_.extend(StickyMethods, {
     this.initializeSticky();
     if (!this.isMobile()) {
       this.selectizeCountry();
+    }
+  },
+
+  handleSuccess: function(e, data) {
+    if (data.follow_up_url) {
+      window.location.href = data.follow_up_url
+    } else {
+      // this should never happen, but just in case.
+      alert("You've signed the petition! Thanks so much!");
     }
   },
 
@@ -63,17 +75,6 @@ const ActionBar = Backbone.View.extend(_.extend(StickyMethods, {
     this.$('.action-bar__main').css('top', `${height}px`);
     this.$el.parent('.sticky-wrapper').css('top', `-${height}px`);
   },
-
-  selectizeCountry: function() {
-    $('.action-bar__country-selector').selectize();
-  },
-
-  clearForm: function(){
-    let $fields_holder = this.$('.form__group--prefilled');
-    $fields_holder.removeClass('form__group--prefilled');
-    $fields_holder.find('input').removeAttr('value');
-    $('.action-bar__welcome-text').addClass('hidden-irrelevant');
-  }
 
 }));
