+++ app/views/plugins/actions/_action.liquid
@@ -22,40 +22,10 @@
         <div class="mobile-hide">
           {% include 'Thermometer' %}
         </div>
-
-        <form action='/api/pages/{{id}}/actions' data-remote='true' method='post' class='form--big action'>
-          {% if member %}
-              <div class="action-bar__welcome-text">
-                Welcome back, {{member.welcome_name}}!
-                <br />
-                <a href="javascript:;" class="action-bar__clear-form">Not you?</a>
-              </div>
-          {% endif %}
-          <input type='hidden' name='form_id' value='{{ plugins.action[ref].form_id }}' />
-          {% for field in plugins.action[ref].fields %}
-            <div class="form__group {% if member[field.name] %}form__group--prefilled{% endif %}">
-              {% case field.data_type %}
-
-              {% when 'text' %}
-                <input type="text" name="{{field.name}}" class="form-control" id="" placeholder="{{ field.label }}" value="{{member[field.name]}}"/>
-              {% when 'email' %}
-                <input type="email" name="{{field.name}}" class="form-control mailcheck" id="" placeholder="{{ field.label }}" value="{{member[field.name]}}" />
-              {% when 'phone' %}
-                <input type="tel" name="{{field.name}}" class="form-control" id="" placeholder="{{ field.label }}" value="{{member[field.name]}}" />
-              {% when 'checkbox' %}
-                <label class="checkbox-label">
-                  <input type="checkbox" name="{{field.name}}" /> {{ field.label }}
-                </label>
-              {% when 'country' %}
-                <select name="{{field.name}}" class="action-bar__country-selector" placeholder="{{field.label}}">
-                  <option></option>
-                  {{ country_option_tags }}
-                </select>
-              {% endcase %}
-            </div>
-          {% endfor %}
-          <button type="submit" class="button action-bar__submit-button">{{ plugins.action[ref].cta }}</button>
-        </form>
+        {% include 'Form',
+          cta: plugins.action[ref].cta,
+          form_id: plugins.action[ref].form_id,
+          fields: plugins.action[ref].fields %}
       </div>
     </div>
   </div>
